use.intermediate resutls for gapfill_landsat

+++ R/gapfill_landsat.R
@@ -1,31 +1,41 @@
 #' Gapfilling for repeated measurements over years
 #'
-#' @param doy 
-#' @param mat 
-#' @param img.nrow 
-#' @param img.ncol 
-#' @param doyeval 
-#' @param h.tcov 
-#' @param h.tsigma2 
-#' @param h.scov 
-#' @param h.ssigma2 
-#' @param nnr 
-#' @param outlier.action 
-#' @param intermediate.save 
-#' @param intermediate.dir 
+#' @param year vecotr of year
+#' @param doy vecotr of doy
+#' @param mat matrix with number of row equals length of doy 
+#' @param img.nrow row number for the image to gapfill
+#' @param img.ncol column number for the image to gapfill
+#' @param doyeval a vector of DOY on which to perform the gapfill
+#' @param h.tcov bandwidth for temporal covariance estimation
+#' @param h.tsigma2 bandwith for temporal variance estimation 
+#' @param h.scov bandwidth for spatial covariance estimation
+#' @param h.ssigma2 bandwidth for spatial variance estimation
+#' @param nnr number of nearest neighbor pixels to use for spatial covariance estimation
+#' @param outlier.action "keep" to keep outliers; "remove" to replace outliers with gapfilled values
+#' @param intermediate.save TRUE or FASLE; whether to save the intermediate results
+#' @param intermediate.dir directory where to save the intermediate results
+#' @param use.intermediate.result whether to use the intermediate results in the 'intermediate.dir' folder instead
+#' of recalcuating 
+#' @param teff TRUE or FALSE, wheter to calculate the temporal effect
+#' @param seff TRUE or FALSE, wheter to calculate the spatial effect
 #'
-#' @return
+#' @return List of length 3 with entries:
+#' \itemize{
+#'   \item imat imputed matrix of `mat`
+#'   \item idx a list of image indexes information
+#'   \item outlier a list of image outliers information
+#' }
 #' @export
 #'
 #' @examples
 gapfill_landsat <- function(year, doy, mat, img.nrow, img.ncol, doyeval = 1:365,  h.tcov = 100, h.tsigma2 = 300,
                             h.scov = 2, h.ssigma2 = 2, nnr = 10, outlier.action = c("keep", "remove"),
-                            intermediate.save = TRUE, intermediate.dir = "./output/", teff = TRUE,
-                            seff = TRUE){
+                            intermediate.save = TRUE, intermediate.dir = "./output/",
+                            use.intermediate.result = TRUE, teff = TRUE, seff = TRUE){
   if(intermediate.save){
     if(!dir.exists(intermediate.dir)){
       cat("Folder 'output' is created to save intermediate results.")
-      dir.create(intermediate.dir)
+      dir.create(intermediate.dir, recursive = TRUE)
     }
   }
   
@@ -33,9 +43,14 @@ gapfill_landsat <- function(year, doy, mat, img.nrow, img.ncol, doyeval = 1:365,
   ###################################
   #### 1. Overall mean estimaton ####
   ###################################
-  meanest = meanEst(doy, mat, doyeval = doyeval)
-  if(intermediate.save)
-    saveRDS(meanest, paste0(intermediate.dir, "meanest.rds"))
+  if(use.intermediate.result & file.exists(paste0(intermediate.dir, "meanest.rds"))){
+    meanest = readRDS(paste0(intermediate.dir, "meanest.rds"))
+  } else {
+    meanest = meanEst(doy, mat, doyeval = doyeval, msk = msk)
+    if(intermediate.save)
+      saveRDS(meanest, paste0(intermediate.dir, "meanest.rds"))
+  }
+  
   ###################################
   #### 2. Time effect estimation ####
   ###################################
@@ -54,26 +69,38 @@ gapfill_landsat <- function(year, doy, mat, img.nrow, img.ncol, doyeval = 1:365,
   ## estimate the temporal effect using residuals
   ## result is a 3d array with the first dimension year, second dimension doy and third dimension pixel index
   if(teff){
-    teffarray = teffEst(year, doy, rmat, doyeval = meanest$doyeval, h.cov = h.tcov, h.sigma2 = h.tsigma2)
-    if(intermediate.save)
-      saveRDS(teffarray, paste0(intermediate.dir, "teffarray.rds"))
-  ######################################
-  #### 3. Spatial effect estimation ####
-  ######################################
-  ## claculate residuals after removing temporal effect
-  yearidx = unlist(lapply(year, function(x,y) which(y == x), y = as.numeric(dimnames(teffarray)[[1]])))
-  doyidx = unlist(lapply(doy, function(x,y) which(y == x), y = as.numeric(dimnames(teffarray)[[2]])))
-  for(i in 1:nrow(rmat)){
-    rmat[i,] = rmat[i,] - teffarray[yearidx[i], doyidx[i],]
-  }
+    if(use.intermediate.result & file.exists(paste0(intermediate.dir, "teffarray.rds"))){
+      teffarray = readRDS(paste0(intermediate.dir, "teffarray.rds"))
+    } else {
+      teffarray = teffEst(year, doy, rmat, doyeval = meanest$doyeval, h.cov = h.tcov, h.sigma2 = h.tsigma2)
+      if(intermediate.save)
+        saveRDS(teffarray, paste0(intermediate.dir, "teffarray.rds"))
+    }
+    
+    
+    ######################################
+    #### 3. Spatial effect estimation ####
+    ######################################
+    ## claculate residuals after removing temporal effect
+    yearidx = unlist(lapply(year, function(x, y)
+      which(y == x), y = as.numeric(dimnames(teffarray)[[1]])))
+    doyidx = unlist(lapply(doy, function(x, y)
+      which(y == x), y = as.numeric(dimnames(teffarray)[[2]])))
+    for (i in 1:nrow(rmat)) {
+      rmat[i, ] = rmat[i, ] - teffarray[yearidx[i], doyidx[i], ]
+    }
   }
   
   ## estimate the spatial effect using residuals
   ## result is a 3d array with the first dimension year, second dimension doy and third dimension pixel index
   if(seff){
-    seffest = seffEst(rmat, img.nrow, img.ncol, nnr = nnr, h.cov = h.scov, h.sigma2 = h.ssigma2)
-    if(intermediate.save)
-      saveRDS(seffest, paste0(intermediate.dir, "seffest.rds"))
+    if(use.intermediate.result & file.exists(paste0(intermediate.dir, "seffest.rds"))){
+      seffest = readRDS(paste0(intermediate.dir, "seffest.rds"))
+    } else {
+      seffest = seffEst(rmat, img.nrow, img.ncol, nnr = nnr, h.cov = h.scov, h.sigma2 = h.ssigma2)
+      if(intermediate.save)
+        saveRDS(seffest, paste0(intermediate.dir, "seffest.rds"))
+    }
   }
   
   #######################
